(feat): operator: add ability to choose an l1 block (#26)
This PR adds ability to choose an L1 block for proof generation. 

Rationale:
The `/staking/validator-set` endpoint of heimdall returns list of validators which has a field called `last_updated` representing a value denoting the recent most L1 block number when the stake was updated for a particular validator. The value is obtained by `block_number * 100000 + tx_log_index`. Hence, on dividing this value by `100k`, we get the block number when a particular validator's stake was updated and we choose the max out of it. 

Because we can't verify all blocks in solidity, we also have to check if this block we derived is in that limit or not. If not, we choose the latest block. The next step is to also check if there was any stake update on L1 between the max block and current block. If there's any, we'll have to choose a number lower that that block instead.