(feat): add prev verified block hash and ensure sequence (#27)
This PR adds `lastVerifiedBorBlockHash` in the verifier contract which is updated when the proof is successfully verified.

This info is later used to ensure that the proof being generated is for a block which is ahead of the previous verified block. Hence, we pass the last verified block as well to the inputs to the prover and assert if the block hash matches with the hash on L1 and if the number of current bor block is greater than the previous one.